requirements.txt
praw==7.7.1
python-telegram-bot==20.7
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0 